more cleanup of old csb toolbar gui code
git-svn-id: https://tuftsvue.svn.sourceforge.net/svnroot/tuftsvue/trunk@1384 d66e5487-c4fa-4dda-8c8e-ca23240363c0

+++ VUE2/src/tufts/vue/LineIcon.java
@@ -30,7 +30,7 @@ public class LineIcon extends BlobIcon
         this( pWidth, pHeight, null, null);
     }
 	
-    public LineIcon( int pWidth,int  pHeight, int pWeight) {
+    public LineIcon( int pWidth,int  pHeight, float pWeight) {
         this( pWidth, pHeight, null, null);
         setWeight(pWeight);
     }
